Change internal storage to std::list
Current implementation makes it dangerous to insert and/or remove tasks
during tick callback is in action: Insert element may cause std::vector
to reallocate, and remove element will shift elements in the container,
both may invalidate the pointers and/or references to elements, which
will lead to errorneous access and/or segmentation fault.

Since the traversal of the container is always sequential, making
the container std::list will solve the problem, as the only time
pointers and/or references are invalidated is when the element is
removed. Also, std::list::sort is always stable, so sorting in the
middle of iteration won't scramble the task waiting to be triggered.

This change allows adding tasks in callback as well as removing tasks
that's not the currently triggering task.